+++ CHANGELOG.md
@@ -21,6 +21,8 @@
 
 - [prql-python] Compilation options can now be specified from Python. (@eitsupi,
   #1807)
+- [prql-js] Default compile target changed from `Sql(Generic)` to `Sql(None)`.
+  (@eitsupi, #1856)
 
 **Internal changes**: